refactoring: change var names, delete unnecessary null-check in ImageAdapter
