+++ .idea/misc.xml
@@ -3,15 +3,11 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Users/alkonov/Documents/GalTest/app/src/main/res/layout/activity_main.xml" value="0.1" />
-        <entry key="..\:/Users/alkonov/Documents/GalTest/app/src/main/res/layout/activity_view.xml" value="0.1" />
-        <entry key="..\:/Users/alkonov/Documents/GalTest/app/src/main/res/layout/gallery_item.xml" value="0.1" />
-        <entry key="..\:/Users/alkonov/Documents/GalTest/app/src/main/res/layout/view_fragment.xml" value="0.12183383991894631" />
+        <entry key="..\:/Users/alkonov/AndroidStudioProjects/GalleryApp/app/src/main/res/layout/activity_image_view.xml" value="0.1671875" />
+        <entry key="..\:/Users/alkonov/AndroidStudioProjects/GalleryApp/app/src/main/res/layout/activity_main.xml" value="0.1671875" />
+        <entry key="..\:/Users/alkonov/AndroidStudioProjects/GalleryApp/app/src/main/res/layout/gallery_item.xml" value="0.1671875" />
       </map>
     </option>
-  </component>
-  <component name="ProjectPlainTextFileTypeManager">
-    <file url="file://$PROJECT_DIR$/app/src/main/res/layout/view_fragment.xml" />
   </component>
   <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
