+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
             android:name=".ImageViewActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name="com.example.galleryapp.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
