+++ app/src/main/java/com/example/galleryapp/ImageAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.gallery
+package com.example.galleryapp
 
 import android.content.Context
 import android.view.LayoutInflater
@@ -6,12 +6,11 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
 import android.widget.ImageView
-import com.example.galtest.Model
 import com.example.galtest.R
 
 class ImageAdapter (
     var itemModel: ArrayList<Model>,
-    var context: Context
+    context: Context
     ) : BaseAdapter() {
 
     var layoutInflater = context.getSystemService(Context.LAYOUT_INFLATER_SERVICE) as LayoutInflater
@@ -20,25 +19,24 @@ class ImageAdapter (
         return itemModel.size
     }
 
-    override fun getItem(p0: Int): Any {
-        return itemModel[p0]
+    override fun getItem(item: Int): Model {
+        return itemModel[item]
     }
 
-    override fun getItemId(p0: Int): Long {
-        return p0.toLong()
+    override fun getItemId(itemId: Int): Long {
+        return itemId.toLong()
     }
 
     override fun getView(position: Int, view: View?, viewGroup: ViewGroup?): View {
 
-        var view = view;
-        if(view == null){
-            view = layoutInflater.inflate(R.layout.gallery_item, viewGroup, false)
+        var localView = view;
+        if(localView == null){
+            localView = layoutInflater.inflate(R.layout.gallery_item, viewGroup, false)
         }
 
-        var imageView = view?.findViewById<ImageView>(R.id.imageView)
-        imageView?.setImageResource(itemModel[position].image!!)
+        val imageView = localView!!.findViewById<ImageView>(R.id.imageView)
+        imageView?.setImageResource(itemModel[position].image)
 
-        return view!!
-
+        return localView
     }
 }
