+++ app/src/main/java/com/example/galleryapp/Model.kt
@@ -1,13 +1,5 @@
-package com.example.galtest
+package com.example.galleryapp
 
 import java.io.Serializable
 
-class Model : Serializable{
-    var name: String? = null
-    var image: Int = 0
-
-    constructor(name:String, image:Int){
-        this.name = name
-        this.image = image
-    }
-}
+data class Model(var name:String, var image:Int) : Serializable
